refactor: extract device.py
