+++ .gitignore
@@ -235,3 +235,5 @@ torch_npu/utils/exposed_api.py
 torch_npu/_op_plugin_docs.py
 
 torch_npu/include
+
+kernel_meta
